+++ src/controllers/account-groups/get-groups.ts
@@ -1,7 +1,7 @@
 import type { CustomResponse } from '@common/types';
 import { API_RESPONSE_STATUS } from 'shared-types';
 import { errorHandler } from '@controllers/helpers';
-import * as accountGroupService from '@services/account-groups/account-groups.service';
+import * as accountGroupService from '@services/account-groups';
 
 export const getAccountGroups = async (req, res: CustomResponse) => {
   try {
