+++ src/controllers/account-groups/remove-account-from-group.ts
@@ -3,7 +3,7 @@ import type { CustomResponse } from '@common/types';
 import { API_RESPONSE_STATUS } from 'shared-types';
 import { recordId } from '@common/lib/zod/custom-types';
 import { errorHandler } from '@controllers/helpers';
-import * as accountGroupService from '@services/account-groups/account-groups.service';
+import * as accountGroupService from '@services/account-groups';
 
 export const removeAccountFromGroup = async (req, res: CustomResponse) => {
   try {
