+++ src/services/account-groups/add-account-to-group.ts
@@ -1,0 +1,14 @@
+import AccountGrouping from '@models/accounts-groups/AccountGrouping.model';
+import { withTransaction } from '../common';
+
+export const addAccountToGroup = withTransaction(
+  async ({
+    accountId,
+    groupId,
+  }: {
+    accountId: number;
+    groupId: number;
+  }): Promise<AccountGrouping> => {
+    return AccountGrouping.create({ accountId, groupId });
+  },
+);
