+++ src/services/account-groups/move-account-between-groups.ts
@@ -1,0 +1,19 @@
+import AccountGroup from '@models/accounts-groups/AccountGroups.model';
+import { withTransaction } from '../common';
+
+export const moveAccountGroup = withTransaction(
+  async ({
+    groupId,
+    newParentGroupId,
+    userId,
+  }: {
+    groupId: number;
+    newParentGroupId: number | null;
+    userId: number;
+  }): Promise<[number, AccountGroup[]]> => {
+    return AccountGroup.update(
+      { parentGroupId: newParentGroupId },
+      { where: { id: groupId, userId }, returning: true },
+    );
+  },
+);
