+++ src/services/account-groups/remove-account-to-group.ts
@@ -1,0 +1,8 @@
+import AccountGrouping from '@models/accounts-groups/AccountGrouping.model';
+import { withTransaction } from '../common';
+
+export const removeAccountFromGroup = withTransaction(
+  async ({ accountId, groupId }: { accountId: number; groupId: number }): Promise<number> => {
+    return AccountGrouping.destroy({ where: { accountId, groupId } });
+  },
+);
